+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     * New: Add `get_filter_names()` as a helper method for getting the default and third-party filter names
     * New: Add `unhook_default_filters()` as a helper method to unhook plugin's default filters hooked to `text_hover()`
     * New: Add test case for hover text that includes HTML
+    * New: Add tests for setting defaults
     * Change: Add `$priority` argument to `test_hover_applies_to_default_filters()`
 
 ## 3.9.1 _(2020-01-12)_